deleted unused files and put Grid in proper folder
